requirements/base.txt
beautifulsoup4==4.5.0
diff-match-patch==20121119
Django==1.9.8
django-import-export==0.5.1
django-iprestrict==1.0.2
django-modelcluster==2.0
django-taggit==0.18.3
django-templatetag-handlebars==1.3.1
django-treebeard==4.0.1
django-user-agents==0.3.0
djangorestframework==3.4.0
GeoIP==1.3.2
geoip2==2.4.0
html5lib==0.999999
maxminddb==1.2.1
Pillow==3.3.0
psycopg2==2.6.2
pycountry==1.20
python-memcached==1.58
pytz==2016.6.1
PyYAML==3.12
requests==2.11.1
six==1.10.0
sqlparse==0.2.2
tablib==0.11.2
ua-parser==0.7.2
Unidecode==0.4.19
user-agents==1.0.1
wagtail==1.6.3
Wand==0.4.3
wheel==0.24.0
Willow==0.3.1